refact separet ReadPackageJson
