--- cmd/completion.go
@@ -1,71 +1,0 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-*/
-package cmd
-
-import (
-	"fmt"
-	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
-	"io"
-	"os"
-)
-
-func NoArgs(cmd *cobra.Command, args []string) error {
-	if len(args) > 0 {
-		return errors.Errorf(
-			"%q accepts no arguments\n\nUsage:  %s",
-			cmd.CommandPath(),
-			cmd.UseLine(),
-		)
-	}
-	return nil
-}
-
-// completionCmd represents the completion command
-
-func newCompletionCmd(out io.Writer) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "completion",
-		Short: "generate autocompletion scripts for the specified shell",
-	}
-	zsh := &cobra.Command{
-		Use:               "zsh",
-		Short:             "generate autocompletion script for zsh",
-		Args:              NoArgs,
-		ValidArgsFunction: noCompletions,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return runCompletionZsh(out, cmd)
-		},
-	}
-	cmd.AddCommand(zsh)
-	return cmd
-}
-
-func runCompletionZsh(out io.Writer, cmd *cobra.Command) error {
-	err := cmd.Root().GenZshCompletion(out)
-
-	// Cobra doesn't source zsh completion file, explicitly doing it here
-	fmt.Fprintf(out, "compdef _jscript jscript")
-
-	return err
-}
-
-// Function to disable file completion
-func noCompletions(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-	return nil, cobra.ShellCompDirectiveNoFileComp
-}
-
-func init() {
-	rootCmd.AddCommand(newCompletionCmd(os.Stdout))
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// completionCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// completionCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
