+++ cmd/root.go
@@ -4,50 +4,23 @@ Copyright © 2023 NAME HERE <EMAIL ADDRESS>
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
-	"os"
-	"strings"
-
+	"github.com/ashigirl96/jscript/pkg"
 	"github.com/spf13/cobra"
+	"os"
 )
 
-type Scripts = map[string]string
-
-type PackageJson struct {
-	Scripts Scripts `json:"scripts"`
-}
-
-func (p *PackageJson) String() string {
-	var result []string
-	for name, command := range p.Scripts {
-		s := fmt.Sprintf("\x1b[32m%s\x1b[m:", name)
-		script := fmt.Sprintf("%-24s%s", s, command)
-		result = append(result, script)
-	}
-	return strings.Join(result, "\n")
-}
-
-var packageJson PackageJson
-
-// rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
 	Use: "jscript",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		packageJsonPath, err := cmd.Flags().GetString("package")
-		if err != nil {
-			return err
-		}
-		raw, err := os.ReadFile(packageJsonPath)
-		if err != nil {
-			return err
-		}
-		err = json.Unmarshal(raw, &packageJson)
-		if err != nil {
-			return err
-		}
-		fmt.Println(&packageJson)
-		return nil
+	//PreRunE: func(cmd *cobra.Command, args []string) error {
+	//	packageJsonPath, err := cmd.Flags().GetString("package")
+	//	if err != nil {
+	//		return err
+	//	}
+	//	return pkg.ReadPackageJson(packageJsonPath)
+	//},
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println(&pkg.PackageJson)
 	},
 }
 
@@ -60,8 +33,8 @@ func Execute() {
 	}
 }
 
-func init() {
-	// Cobra also supports local flags, which will only run
-	// when this action is called directly.
-	rootCmd.Flags().StringP("package", "p", "./package.json", "path of package.json")
-}
+//func init() {
+//	// Cobra also supports local flags, which will only run
+//	// when this action is called directly.
+//	rootCmd.Flags().StringP("package", "p", "./package.json", "path of package.json")
+//}
