+++ cmd/run.go
@@ -1,0 +1,24 @@
+/*
+Copyright © 2023 NAME HERE <EMAIL ADDRESS>
+*/
+package cmd
+
+import (
+	"fmt"
+	"github.com/ashigirl96/jscript/pkg"
+
+	"github.com/spf13/cobra"
+)
+
+// runCmd represents the run command
+var runCmd = &cobra.Command{
+	Use:   "run",
+	Short: "A brief description of your command",
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println(&pkg.PackageJson)
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(runCmd)
+}
